:truck: Moved UDPConnection class to .hpp file
